Writer queue is filled based on the files in temp and not based on the crawler
