+++ crawler/oldDatabaseWriter.py
@@ -54,7 +54,6 @@ class Writer:
         user_id = file_name[5:].split('_')[0]
         if file_name[-8:-4] == 'save':
             self.__log_to_file("WRITING: %s" % user_id)
-            self.__delete_old_relations(user_id)
             self.__write_followers(user_id)
         elif file_name[-10:-4] == 'delete':
             self.__delete_user(user_id)
@@ -70,21 +69,52 @@ class Writer:
         return [d['followerid'] for d in self.__run_get_query(query).data()]
 
     def __write_followers(self, user_id):
-        num_new_followers = 0
-        batch_size = 50000
-        self.__log_to_file("Writing new followers for user %s." % user_id)
-        with open("temp/%s_save.txt" % user_id, "r") as followers_file:
-            batch = []
-            for line in followers_file.readlines():
-                batch.extend(line.replace('\n', '').split(','))
-                if len(batch) >= batch_size:
-                    self.__save_user_followers(user_id, batch)
-                    num_new_followers += len(batch)
-                    batch = []
-            if len(batch) > 0:
-                self.__save_user_followers(user_id, batch)
-                num_new_followers += len(batch)
-        self.__update_user(user_id, num_new_followers)
+        num_followers = 0
+        batch = 0
+        batch_size = 1000000
+        self.__log_to_file("Requesting followers...")
+        followers_in_db = []
+        next_batch = self.__request_followers(user_id, batch, batch_size)
+        while len(next_batch) > 0:
+            t0 = time.time()
+            followers_in_db.extend(next_batch)
+            batch += 1
+            next_batch = self.__request_followers(user_id, batch, batch_size)
+            self.__log_to_file("Time to complete batch %s: %s seconds." % (batch, time.time() - t0))
+        self.__log_to_file("Followers requested successfully! List has %s followers." % len(followers_in_db))
+        list_len = len(followers_in_db)
+        if list_len > 0:
+            with open("temp/%s_save.txt" % user_id, "r") as followers_file:
+                followers_in_db = set(followers_in_db)
+                set_len = len(followers_in_db)
+                self.__log_to_file("List converted successfully into set! Set has %s followers." % set_len)
+                if list_len != set_len:
+                    self.__log_to_file("WARNING!!!WARNING!!!WARNING!!!WARNING!!!WARNING!!!WARNING!!!\n" +
+                                       "      OLD FOLLOWERS LIST AND SET GOT DIFFERENT LENGTHS\n" +
+                                       "WARNING!!!WARNING!!!WARNING!!!WARNING!!!WARNING!!!WARNING!!!")
+                for line in followers_file.readlines():
+                    new_followers = [f for f in line.replace('\n', '').split(',') if f not in followers_in_db]
+                    old_followers = [f for f in line.replace('\n', '').split(',') if f in followers_in_db]
+                    self.__log_to_file("Removing %s followers from set" % len(old_followers))
+                    for old in old_followers:
+                        followers_in_db.remove(old)
+                    self.__log_to_file("Set has, now, %s users" % len(followers_in_db))
+                    batch_num_followers = len(new_followers)
+                    if batch_num_followers > 0:
+                        self.__save_user_followers(user_id, new_followers)
+                    num_followers += batch_num_followers
+                self.__delete_relations(user_id, list(followers_in_db))
+                self.__log_to_file("Old relations deleted: %s" % len(followers_in_db))
+        else:
+            self.__log_to_file("No user was in the database yet...")
+            with open("temp/%s_save.txt" % user_id, "r") as followers_file:
+                for line in followers_file.readlines():
+                    followers = line.replace('\n', '').split(',')
+                    batch_num_followers = len(followers)
+                    if batch_num_followers > 0:
+                        self.__save_user_followers(user_id, followers)
+                    num_followers += batch_num_followers
+        self.__update_user(user_id, num_followers)
 
     def __save_user_followers(self, user_id, followers):
         # Add a batch of followers to the db
@@ -114,20 +144,17 @@ class Writer:
         self.__run_query(query)
         self.__log_to_file("Deleted user %s\n\n\n" % user_id)
 
-    def __delete_old_relations(self, user_id):
-        self.__log_to_file("Deleting old relations from user %s." % user_id)
-        batch_size = 100000
-        while True:
-            query = "MATCH (u:QUEUED:USER{uid:'%s'})<-[r:FOLLOWS]-() " % user_id
-            query += "WITH r LIMIT %s " % batch_size
-            query += "DELETE r "
-            query += "RETURN COUNT(r)"
-            data = self.__run_get_query(query).data()[0]['COUNT(r)']
-            self.__log_to_file("%s relations deleted." % data)
-            if data < batch_size:
-                break
-            else:
-                continue
+    def __delete_relations(self, user_id, old_followers):
+        self.__log_to_file("Deleting %s invalid relations from user %s." % (len(old_followers), user_id))
+        rest = batch = old_followers
+        batch_size = 5000
+        while len(batch) > 0:
+            batch, rest = rest[:batch_size], rest[batch_size:]
+            query = "WITH %s AS oldfollowers " % batch
+            query += "MATCH (u:QUEUED:USER{uid:'%s'})<-[r:FOLLOWS]-(f:USER) WHERE f.uid IN oldfollowers " % user_id
+            query += "DELETE r"
+            self.__run_query(query)
+            self.__log_to_file("Batch of %s old followers deleted." % len(batch))
 
     def __skip_user(self, user_id):
         # skipping users having the wrong language
